Move playlist files to their own directory.
