+++ server/server.js
@@ -31,8 +31,8 @@ import {Control} from './control.js';
 import {ModuleLoader} from './modules/module_loader.js';
 import {ServerModulePlayer} from './modules/server_module_player.js';
 import peer from 'peer';
-import {PlaylistDriver} from './modules/playlist_driver.js';
-import {PlaylistLoader} from './modules/playlist_loader.js';
+import {PlaylistDriver} from './playlist/playlist_driver.js';
+import {PlaylistLoader} from './playlist/playlist_loader.js';
 import {makeConsoleLogger} from '../lib/console_logger.js';
 import {captureLog} from './util/last_n_errors_logger.js';
 import {addLogger, easyLog} from '../lib/log.js';
